feat(core): provide resolveConfig, used to merge user config
provide the way to merge user config with default resolveConfig

BREAKING CHANGE: use resolveConfig in tailwind.config.js to merge @retail/ui default
