+++ packages/core/package.json
@@ -40,9 +40,13 @@
     "lint": "tsdx lint"
   },
   "peerDependencies": {
-    "react": ">=16"
+    "react": ">=16",
+    "tailwindcss": ">=1"
   },
   "dependencies": {
+    "deepmerge": "^4.2.2",
+    "@tailwindcss/custom-forms": "0.2.1",
+    "tailwindcss-multi-theme": "1.0.3",
     "@retail-ui/action-buttons": "^0.3.1",
     "@retail-ui/alert": "^0.3.1",
     "@retail-ui/avatar": "^0.2.0",
