+++ packages/core/src/index.tsx
@@ -11,3 +11,4 @@ export * from '@retail-ui/hooks'
 export * from '@retail-ui/icon'
 export * from '@retail-ui/spinner'
 export * from '@retail-ui/theme'
+export * from './resolveConfig'
